CLN: move pytables tests to tests/io/pytables dir (#26986)
