+++ pandas/tests/io/pytables/test_pytables.py
