+++ pandas/tests/io/pytables/test_pytables_missing.py
